\name{exforParser-package}
\alias{exforParser-package}
\alias{exforParser}
\docType{package}
\title{
  exforParser - Extract the Information from EXFOR entries
}
\description{
  EXFOR entries given as character strings can be 
  converted into nested lists which can be conveniently
  handled in R. It is also possible to convert 
  EXFOR entries to JSON objects which can be exported
  to a variety of other high-level languages. 
}
\author{
Georg Schnabel

Maintainer: Georg Schnabel <georg.schnabel@nucleardata.com>
}
\keyword{ package }
